refactor(replay): rename offline-viz into replay
